alx_travel_app/requirements.txt
django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
celery==5.5.3
requests==2.32.5
python-dotenv==1.1.1
kombu==5.3.4
amqp==5.2.0